[Store] Split snapshot MasterService unit tests
