+++ mooncake-store/tests/ha/snapshot/master_service/retention_test.cpp
@@ -1,0 +1,782 @@
+#include "fixture.h"
+
+namespace mooncake::test {
+TEST_F(MasterServiceSnapshotTest, CleanupStaleHandlesTest) {
+    service_.reset(new MasterService());
+
+    // Mount a segment for testing
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 16;  // 16MB
+    auto segment = MakeSegment("test_segment", buffer, size);
+    UUID client_id = generate_uuid();
+
+    // Mount the segment
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    // Create an object that will be stored in the segment
+    std::string key = "segment_object";
+    uint64_t slice_length = 1024 * 1024;  // One 1MB slice
+    ReplicateConfig config;
+    config.replica_num = 1;  // One replica
+
+    // Create the object
+    auto put_start_result =
+        service_->PutStart(client_id, key, "default", slice_length, config);
+    ASSERT_TRUE(put_start_result.has_value());
+    auto put_end_result =
+        service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+    ASSERT_TRUE(put_end_result.has_value());
+
+    // Verify object exists
+    auto get_result = service_->GetReplicaList(key, "default");
+    ASSERT_TRUE(get_result.has_value());
+    auto retrieved_replicas = get_result.value().replicas;
+    ASSERT_EQ(1, retrieved_replicas.size());
+
+    // Unmount the segment
+    auto unmount_result1 = service_->UnmountSegment(segment.id, client_id);
+    ASSERT_TRUE(unmount_result1.has_value());
+
+    // Try to get the object - it should be automatically removed since the
+    // replica is invalid
+    auto get_result2 = service_->GetReplicaList(key, "default");
+    EXPECT_FALSE(get_result2.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_NOT_FOUND, get_result2.error());
+
+    // Mount the segment again
+    mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    // Create another object
+    std::string key2 = "another_segment_object";
+    auto put_start_result2 =
+        service_->PutStart(client_id, key2, "default", slice_length, config);
+    ASSERT_TRUE(put_start_result2.has_value());
+    auto put_end_result2 =
+        service_->PutEnd(client_id, key2, "default", ReplicaType::MEMORY);
+    ASSERT_TRUE(put_end_result2.has_value());
+
+    // Verify we can get it
+    auto get_result3 = service_->GetReplicaList(key2, "default");
+    ASSERT_TRUE(get_result3.has_value());
+
+    // Unmount the segment
+    auto unmount_result2 = service_->UnmountSegment(segment.id, client_id);
+    ASSERT_TRUE(unmount_result2.has_value());
+
+    // Try to remove the object that should already be cleaned up
+    auto remove_result = service_->Remove(key2, "default");
+    EXPECT_FALSE(remove_result.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_NOT_FOUND, remove_result.error());
+}
+
+TEST_F(MasterServiceSnapshotTest, RemoveLeasedObject) {
+    const uint64_t kv_lease_ttl = 50;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    std::string key = "test_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    // Verify lease is granted on ExistsKey
+    auto put_start_result =
+        service_->PutStart(client_id, key, "default", slice_length, config);
+    ASSERT_TRUE(put_start_result.has_value());
+    auto put_end_result =
+        service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+    ASSERT_TRUE(put_end_result.has_value());
+    auto exist_result = service_->ExistKey(key, "default");
+    ASSERT_TRUE(exist_result.has_value());
+    auto remove_result = service_->Remove(key, "default");
+    EXPECT_FALSE(remove_result.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_result.error());
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    auto remove_result2 = service_->Remove(key, "default");
+    EXPECT_TRUE(remove_result2.has_value());
+
+    // Verify lease is extended on successive ExistsKey
+    auto put_start_result2 =
+        service_->PutStart(client_id, key, "default", slice_length, config);
+    ASSERT_TRUE(put_start_result2.has_value());
+    auto put_end_result2 =
+        service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+    ASSERT_TRUE(put_end_result2.has_value());
+    auto exist_result2 = service_->ExistKey(key, "default");
+    ASSERT_TRUE(exist_result2.has_value());
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    auto exist_result3 = service_->ExistKey(key, "default");
+    ASSERT_TRUE(exist_result3.has_value());
+    auto remove_result3 = service_->Remove(key, "default");
+    EXPECT_FALSE(remove_result3.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_result3.error());
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    auto remove_result4 = service_->Remove(key, "default");
+    EXPECT_TRUE(remove_result4.has_value());
+
+    // Verify lease is granted on GetReplicaList
+    auto put_start_result3 =
+        service_->PutStart(client_id, key, "default", slice_length, config);
+    ASSERT_TRUE(put_start_result3.has_value());
+    auto put_end_result3 =
+        service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+    ASSERT_TRUE(put_end_result3.has_value());
+    auto get_result = service_->GetReplicaList(key, "default");
+    ASSERT_TRUE(get_result.has_value());
+    auto remove_result5 = service_->Remove(key, "default");
+    EXPECT_FALSE(remove_result5.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_result5.error());
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    auto remove_result6 = service_->Remove(key, "default");
+    EXPECT_TRUE(remove_result6.has_value());
+
+    // Verify lease is extended on successive GetReplicaList
+    auto put_start_result4 =
+        service_->PutStart(client_id, key, "default", slice_length, config);
+    ASSERT_TRUE(put_start_result4.has_value());
+    auto put_end_result4 =
+        service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+    ASSERT_TRUE(put_end_result4.has_value());
+    auto get_result2 = service_->GetReplicaList(key, "default");
+    ASSERT_TRUE(get_result2.has_value());
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    auto get_result3 = service_->GetReplicaList(key, "default");
+    ASSERT_TRUE(get_result3.has_value());
+    auto remove_result7 = service_->Remove(key, "default");
+    EXPECT_FALSE(remove_result7.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_result7.error());
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    auto remove_result8 = service_->Remove(key, "default");
+    EXPECT_TRUE(remove_result8.has_value());
+
+    // Verify object is removed
+    auto get_result4 = service_->GetReplicaList(key, "default");
+    EXPECT_FALSE(get_result4.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_NOT_FOUND, get_result4.error());
+}
+
+TEST_F(MasterServiceSnapshotTest, RemoveAllLeasedObject) {
+    const uint64_t kv_lease_ttl = 50;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+    for (int i = 0; i < 10; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        uint64_t slice_length = 1024;
+        ReplicateConfig config;
+        config.replica_num = 1;
+        auto put_start_result =
+            service_->PutStart(client_id, key, "default", slice_length, config);
+        ASSERT_TRUE(put_start_result.has_value());
+        auto put_end_result =
+            service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+        ASSERT_TRUE(put_end_result.has_value());
+        if (i >= 5) {
+            auto exist_result = service_->ExistKey(key, "default");
+            ASSERT_TRUE(exist_result.has_value());
+        }
+    }
+    // [Commented for snapshot test] The following RemoveAll would clear data
+    // before TearDown snapshot verification ASSERT_EQ(5,
+    // service_->RemoveAll()); for (int i = 0; i < 5; ++i) {
+    //     std::string key = "test_key" + std::to_string(i);
+    //     auto exist_result = service_->ExistKey(key, "default");
+    //     ASSERT_FALSE(exist_result.value());
+    // }
+    // // wait for all the lease to expire
+    // std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    // ASSERT_EQ(5, service_->RemoveAll());
+    // for (int i = 5; i < 10; ++i) {
+    //     std::string key = "test_key" + std::to_string(i);
+    //     auto exist_result = service_->ExistKey(key, "default");
+    //     ASSERT_FALSE(exist_result.value());
+    // }
+}
+
+TEST_F(MasterServiceSnapshotTest, EvictObject) {
+    // set a large kv_lease_ttl so the granted lease will not quickly expire
+    const uint64_t kv_lease_ttl = 2000;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+    // Mount a segment that can hold about 1024 * 16 objects.
+    // As the eviction is processed separately for each shard,
+    // we need to fill each shard with enough objects to thoroughly
+    // test the eviction process.
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 16 * 15;
+    constexpr size_t object_size = 1024 * 15;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, size);
+
+    // Verify if we can put objects more than the segment can hold
+    int success_puts = 0;
+    for (int i = 0; i < 1024 * 16 + 50; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        uint64_t slice_length = object_size;
+        ReplicateConfig config;
+        config.replica_num = 1;
+        auto put_start_result =
+            service_->PutStart(client_id, key, "default", slice_length, config);
+        if (put_start_result.has_value()) {
+            auto put_end_result = service_->PutEnd(client_id, key, "default",
+                                                   ReplicaType::MEMORY);
+            ASSERT_TRUE(put_end_result.has_value());
+            success_puts++;
+        } else {
+            // wait for eviction to work
+            std::this_thread::sleep_for(std::chrono::milliseconds(50));
+        }
+    }
+    ASSERT_GT(success_puts, 1024 * 16);
+    // [Commented for snapshot test] The following RemoveAll would clear data
+    // before TearDown snapshot verification
+    // std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    // service_->RemoveAll();
+
+    // Wait for eviction to stabilize before snapshot in TearDown
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl + 50));
+}
+
+TEST_F(MasterServiceSnapshotTest, TryEvictLeasedObject) {
+    // set a large kv_lease_ttl so the granted lease will not quickly expire
+    const uint64_t kv_lease_ttl = 500;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 16;
+    constexpr size_t object_size = 1024 * 1024;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, size);
+
+    // Verify leased object will not be evicted.
+    int success_puts = 0;
+    int failed_puts = 0;
+    std::vector<std::string> leased_keys;
+    for (int i = 0; i < 16 + 10; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        uint64_t slice_length = object_size;
+        ReplicateConfig config;
+        config.replica_num = 1;
+        auto put_start_result =
+            service_->PutStart(client_id, key, "default", slice_length, config);
+        if (put_start_result.has_value()) {
+            auto put_end_result = service_->PutEnd(client_id, key, "default",
+                                                   ReplicaType::MEMORY);
+            ASSERT_TRUE(put_end_result.has_value());
+            // the object is leased
+            auto get_result = service_->GetReplicaList(key, "default");
+            ASSERT_TRUE(get_result.has_value());
+            leased_keys.push_back(key);
+            success_puts++;
+        } else {
+            failed_puts++;
+        }
+    }
+    ASSERT_GT(success_puts, 0);
+    ASSERT_GT(failed_puts, 0);
+    // wait for eviction to do eviction
+    std::this_thread::sleep_for(std::chrono::milliseconds(50));
+    // All leased objects should be accessible
+    for (const auto& key : leased_keys) {
+        auto get_result = service_->GetReplicaList(key, "default");
+        ASSERT_TRUE(get_result.has_value());
+    }
+    // [Commented for snapshot test] The following RemoveAll would clear data
+    // before TearDown snapshot verification
+    // std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    // service_->RemoveAll();
+}
+
+TEST_F(MasterServiceSnapshotTest, RemoveSoftPinObject) {
+    const uint64_t kv_lease_ttl = 200;
+    // set a large soft_pin_ttl so the granted soft pin will not quickly expire
+    const uint64_t kv_soft_pin_ttl = 10000;
+    const bool allow_evict_soft_pinned_objects = true;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .set_default_kv_soft_pin_ttl(kv_soft_pin_ttl)
+                              .set_allow_evict_soft_pinned_objects(
+                                  allow_evict_soft_pinned_objects)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+    // Mount segment and put an object
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 16;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, size);
+
+    std::string key = "test_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+    config.with_soft_pin = true;
+
+    // Verify soft pin does not block remove
+    ASSERT_TRUE(
+        service_->PutStart(client_id, key, "default", slice_length, config)
+            .has_value());
+    ASSERT_TRUE(service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY)
+                    .has_value());
+    EXPECT_TRUE(service_->Remove(key, "default").has_value());
+
+    // [Commented for snapshot test] The following RemoveAll would clear data
+    // before TearDown snapshot verification
+    // // Verify soft pin does not block RemoveAll
+    // ASSERT_TRUE(
+    //     service_->PutStart(client_id, key, "default", slice_length, //
+    //     config).has_value());
+    // ASSERT_TRUE(
+    //     service_->PutEnd(client_id, key, "default",
+    //     ReplicaType::MEMORY).has_value());
+    // EXPECT_EQ(1, service_->RemoveAll());
+}
+
+TEST_F(MasterServiceSnapshotTest, SoftPinObjectsNotEvictedBeforeOtherObjects) {
+    const uint64_t kv_lease_ttl = 200;
+    // set a large soft_pin_ttl so the granted soft pin will not quickly expire
+    const uint64_t kv_soft_pin_ttl = 10000;
+    const double eviction_ratio = 0.5;
+    const bool allow_evict_soft_pinned_objects = true;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .set_default_kv_soft_pin_ttl(kv_soft_pin_ttl)
+                              .set_allow_evict_soft_pinned_objects(
+                                  allow_evict_soft_pinned_objects)
+                              .set_eviction_ratio(eviction_ratio)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+
+    // Mount segment and put an object
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t segment_size = 1024 * 1024 * 16;
+    constexpr size_t value_size = 1024 * 1024;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, segment_size);
+
+    // The eviction has random factors, so test 5 times
+    for (int test_i = 0; test_i < 5; test_i++) {
+        // Put pin_key first
+        for (int i = 0; i < 2; i++) {
+            std::string pin_key = "pin_key" + std::to_string(i);
+            uint64_t slice_length = value_size;
+            ReplicateConfig soft_pin_config;
+            soft_pin_config.replica_num = 1;
+            soft_pin_config.with_soft_pin = true;
+
+            ASSERT_TRUE(service_
+                            ->PutStart(client_id, pin_key, "default",
+                                       slice_length, soft_pin_config)
+                            .has_value());
+            ASSERT_TRUE(
+                service_
+                    ->PutEnd(client_id, pin_key, "default", ReplicaType::MEMORY)
+                    .has_value());
+        }
+
+        // Fill the segment to trigger eviction
+        int failed_puts = 0;
+        for (int i = 0; i < 20; i++) {
+            std::string key = "key" + std::to_string(i);
+            uint64_t slice_length = value_size;
+            ReplicateConfig config;
+            config.replica_num = 1;
+            if (service_
+                    ->PutStart(client_id, key, "default", slice_length, config)
+                    .has_value()) {
+                ASSERT_TRUE(
+                    service_
+                        ->PutEnd(client_id, key, "default", ReplicaType::MEMORY)
+                        .has_value());
+            } else {
+                failed_puts++;
+            }
+        }
+        ASSERT_GT(failed_puts, 0);
+        // wait for eviction to do eviction
+        std::this_thread::sleep_for(
+            std::chrono::milliseconds(kv_lease_ttl + 1000));
+        // pin_key should still be accessible
+        for (int i = 0; i < 2; i++) {
+            std::string pin_key = "pin_key" + std::to_string(i);
+            ASSERT_TRUE(
+                service_->GetReplicaList(pin_key, "default").has_value());
+        }
+
+        // wait for the lease to expire
+        std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+        // [Commented for snapshot test] The following RemoveAll would clear
+        // data before TearDown snapshot verification
+        // Only remove all objects before the next turn (skip last round)
+        if (test_i < 4) {
+            service_->RemoveAll();
+        }
+    }
+}
+
+TEST_F(MasterServiceSnapshotTest, SoftPinObjectsCanBeEvicted) {
+    const uint64_t kv_lease_ttl = 200;
+    // set a large soft_pin_ttl so the granted soft pin will not quickly expire
+    const uint64_t kv_soft_pin_ttl = 10000;
+    const bool allow_evict_soft_pinned_objects = true;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .set_default_kv_soft_pin_ttl(kv_soft_pin_ttl)
+                              .set_allow_evict_soft_pinned_objects(
+                                  allow_evict_soft_pinned_objects)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+
+    // Mount segment and put an object
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t segment_size = 1024 * 1024 * 16;
+    constexpr size_t value_size = 1024 * 1024;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, segment_size);
+
+    // Verify if we can put objects more than the segment can hold
+    int success_puts = 0;
+    for (int i = 0; i < 16 + 50; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        uint64_t slice_length = value_size;
+        ReplicateConfig config;
+        config.replica_num = 1;
+        config.with_soft_pin = true;
+        if (service_->PutStart(client_id, key, "default", slice_length, config)
+                .has_value()) {
+            ASSERT_TRUE(
+                service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY)
+                    .has_value());
+            success_puts++;
+        } else {
+            // wait for eviction to work
+            std::this_thread::sleep_for(std::chrono::milliseconds(50));
+        }
+    }
+    ASSERT_GT(success_puts, 16);
+
+    // Wait for eviction to stabilize before TearDown snapshot verification
+    // This ensures no background eviction occurs during snapshot/restore
+    // comparison
+    std::this_thread::sleep_for(std::chrono::milliseconds(500));
+
+    // [Commented for snapshot test] The following RemoveAll would clear data
+    // before TearDown snapshot verification
+    // std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    // service_->RemoveAll();
+}
+
+TEST_F(MasterServiceSnapshotTest, SoftPinExtendedOnGet) {
+    const uint64_t kv_lease_ttl = 200;
+    // The soft pin ttl shall not be too large, otherwise the test will take too
+    // long
+    const uint64_t kv_soft_pin_ttl = 1000;
+    static_assert(
+        kv_soft_pin_ttl > kv_lease_ttl,
+        "kv_soft_pin_ttl must be larger than kv_lease_ttl in this test");
+    const double eviction_ratio = 0.5;
+    const bool allow_evict_soft_pinned_objects = true;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .set_default_kv_soft_pin_ttl(kv_soft_pin_ttl)
+                              .set_allow_evict_soft_pinned_objects(
+                                  allow_evict_soft_pinned_objects)
+                              .set_eviction_ratio(eviction_ratio)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+
+    // Mount segment and put an object
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t segment_size = 1024 * 1024 * 16;
+    constexpr size_t value_size = 1024 * 1024;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, segment_size);
+
+    // The eviction has random factors, so test 3 times
+    for (int test_i = 0; test_i < 3; test_i++) {
+        // Put pin_key first
+        for (int i = 0; i < 2; i++) {
+            std::string pin_key = "pin_key" + std::to_string(i);
+            uint64_t slice_length = value_size;
+            ReplicateConfig soft_pin_config;
+            soft_pin_config.replica_num = 1;
+            soft_pin_config.with_soft_pin = true;
+
+            ASSERT_TRUE(service_->PutStart(client_id, pin_key, "default",
+                                           slice_length, soft_pin_config));
+            ASSERT_TRUE(
+                service_
+                    ->PutEnd(client_id, pin_key, "default", ReplicaType::MEMORY)
+                    .has_value());
+        }
+
+        // Wait for the soft pin to expire
+        std::this_thread::sleep_for(std::chrono::milliseconds(kv_soft_pin_ttl));
+
+        // Get the pin_key to extend the soft pin
+        for (int i = 0; i < 2; i++) {
+            std::string pin_key = "pin_key" + std::to_string(i);
+            ASSERT_TRUE(
+                service_->GetReplicaList(pin_key, "default").has_value());
+        }
+
+        // Fill the segment to trigger eviction
+        int failed_puts = 0;
+        for (int i = 0; i < 16; i++) {
+            std::string key = "key" + std::to_string(i);
+            uint64_t slice_length = value_size;
+            ReplicateConfig config;
+            config.replica_num = 1;
+            if (service_
+                    ->PutStart(client_id, key, "default", slice_length, config)
+                    .has_value()) {
+                ASSERT_TRUE(
+                    service_
+                        ->PutEnd(client_id, key, "default", ReplicaType::MEMORY)
+                        .has_value());
+            } else {
+                failed_puts++;
+            }
+        }
+        ASSERT_GT(failed_puts, 0);
+
+        // wait for eviction
+        std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+
+        // pin_key should still be accessible
+        for (int i = 0; i < 2; i++) {
+            std::string pin_key = "pin_key" + std::to_string(i);
+            ASSERT_TRUE(
+                service_->GetReplicaList(pin_key, "default").has_value());
+        }
+        // [Commented for snapshot test] The following RemoveAll would clear
+        // data before TearDown snapshot verification Only remove all objects
+        // before the next turn (skip last round)
+        if (test_i < 2) {
+            std::this_thread::sleep_for(
+                std::chrono::milliseconds(kv_lease_ttl));
+            service_->RemoveAll();
+        }
+    }
+}
+
+TEST_F(MasterServiceSnapshotTest, SoftPinObjectsNotAllowEvict) {
+    const uint64_t kv_lease_ttl = 200;
+    // set a large soft_pin_ttl so the granted soft pin will not quickly expire
+    const uint64_t kv_soft_pin_ttl = 10000;
+    // set allow_evict_soft_pinned_objects to false to disable eviction of soft
+    // pinned objects
+    const bool allow_evict_soft_pinned_objects = false;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .set_default_kv_soft_pin_ttl(kv_soft_pin_ttl)
+                              .set_allow_evict_soft_pinned_objects(
+                                  allow_evict_soft_pinned_objects)
+                              .build();
+    service_.reset(new MasterService(service_config));
+    const UUID client_id = generate_uuid();
+
+    // Mount segment and put an object
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t segment_size = 1024 * 1024 * 16;
+    constexpr size_t value_size = 1024 * 1024;
+    [[maybe_unused]] const auto context =
+        PrepareSimpleSegment(*service_, "test_segment", buffer, segment_size);
+
+    // Put objects more than the segment can hold
+    std::vector<std::string> success_keys;
+    for (int i = 0; i < 16 + 50; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        uint64_t slice_length = value_size;
+        ReplicateConfig config;
+        config.replica_num = 1;
+        config.with_soft_pin = true;
+        if (service_->PutStart(client_id, key, "default", slice_length, config)
+                .has_value()) {
+            ASSERT_TRUE(
+                service_->PutEnd(client_id, key, "default", ReplicaType::MEMORY)
+                    .has_value());
+            success_keys.push_back(key);
+        } else {
+            // wait for eviction to work
+            std::this_thread::sleep_for(std::chrono::milliseconds(50));
+        }
+    }
+    ASSERT_LE(success_keys.size(), 17);
+    // All soft pinned objects should be accessible
+    for (const auto& key : success_keys) {
+        ASSERT_TRUE(service_->GetReplicaList(key, "default").has_value());
+    }
+    // [Commented for snapshot test] The following RemoveAll would clear data
+    // before TearDown snapshot verification
+    // std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    // service_->RemoveAll();
+}
+
+TEST_F(MasterServiceSnapshotTest, PutStartExpiringTest) {
+    // Reset storage space metrics.
+    MasterMetricManager::instance().reset_allocated_mem_size();
+    MasterMetricManager::instance().reset_total_mem_capacity();
+
+    MasterServiceConfig master_config;
+    master_config.put_start_discard_timeout_sec = 3;
+    master_config.put_start_release_timeout_sec = 5;
+    service_.reset(new MasterService(master_config));
+
+    constexpr size_t kReplicaCnt = 3;
+    constexpr size_t kBaseAddr = 0x300000000;
+    constexpr size_t kSegmentSize = 1024 * 1024 * 16;  // 16MB
+
+    // Mount 3 segments.
+    std::vector<MountedSegmentContext> contexts;
+    contexts.reserve(kReplicaCnt);
+    for (size_t i = 0; i < kReplicaCnt; ++i) {
+        auto context = PrepareSimpleSegment(
+            *service_, "segment_" + std::to_string(i),
+            kBaseAddr + static_cast<size_t>(i) * kSegmentSize, kSegmentSize);
+        contexts.push_back(context);
+    }
+
+    // The client_id used to put objects.
+    auto client_id = generate_uuid();
+    std::string key_1 = "test_key_1", key_2 = "test_key_2";
+    uint64_t value_length = 6 * 1024 * 1024;  // 6MB
+    uint64_t slice_length = value_length;
+    ReplicateConfig config;
+    config.replica_num = kReplicaCnt;
+
+    // Put key_1, should success.
+    auto put_start_result =
+        service_->PutStart(client_id, key_1, "default", slice_length, config);
+    EXPECT_TRUE(put_start_result.has_value());
+    replica_list = put_start_result.value();
+    EXPECT_EQ(replica_list.size(), kReplicaCnt);
+    for (size_t i = 0; i < kReplicaCnt; i++) {
+        EXPECT_EQ(ReplicaStatus::PROCESSING, replica_list[i].status);
+    }
+
+    // Put key_1 again, should fail because the key exists.
+    put_start_result =
+        service_->PutStart(client_id, key_1, "default", slice_length, config);
+    EXPECT_FALSE(put_start_result.has_value());
+    EXPECT_EQ(put_start_result.error(), ErrorCode::OBJECT_ALREADY_EXISTS);
+
+    // Wait for a while until the put-start expired.
+    for (size_t i = 0; i <= master_config.put_start_discard_timeout_sec; i++) {
+        for (auto& context : contexts) {
+            auto result = service_->Ping(context.client_id);
+            EXPECT_TRUE(result.has_value());
+        }
+        std::this_thread::sleep_for(std::chrono::seconds(1));
+    }
+
+    // Put key_1 again, should success because the old one has expired and will
+    // be discarded by this put.
+    put_start_result =
+        service_->PutStart(client_id, key_1, "default", slice_length, config);
+    EXPECT_TRUE(put_start_result.has_value());
+    replica_list = put_start_result.value();
+    EXPECT_EQ(replica_list.size(), kReplicaCnt);
+    for (size_t i = 0; i < kReplicaCnt; i++) {
+        EXPECT_EQ(ReplicaStatus::PROCESSING, replica_list[i].status);
+    }
+
+    // Complete key_1.
+    auto put_end_result =
+        service_->PutEnd(client_id, key_1, "default", ReplicaType::MEMORY);
+    EXPECT_TRUE(put_end_result.has_value());
+
+    // Protect key_1 from eviction.
+    auto get_result = service_->GetReplicaList(key_1, "default");
+    EXPECT_TRUE(get_result.has_value());
+
+    // Put key_2, should fail because the key_1 occupied 12MB (6MB processing,
+    // 6MB discarded but not yet released) on each segment.
+    put_start_result =
+        service_->PutStart(client_id, key_2, "default", slice_length, config);
+    EXPECT_FALSE(put_start_result.has_value());
+    EXPECT_EQ(put_start_result.error(), ErrorCode::NO_AVAILABLE_HANDLE);
+
+    // Wait for a while until the discarded replicas are released.
+    for (size_t i = 0; i <= master_config.put_start_release_timeout_sec -
+                                master_config.put_start_discard_timeout_sec;
+         i++) {
+        for (auto& context : contexts) {
+            auto result = service_->Ping(context.client_id);
+            EXPECT_TRUE(result.has_value());
+        }
+        // Protect key_1 from eviction.
+        auto get_result = service_->GetReplicaList(key_1, "default");
+        EXPECT_TRUE(get_result.has_value());
+        std::this_thread::sleep_for(std::chrono::seconds(1));
+    }
+
+    // Put key_2 again, should success because the discarded replica has been
+    // released.
+    put_start_result =
+        service_->PutStart(client_id, key_2, "default", slice_length, config);
+    EXPECT_TRUE(put_start_result.has_value());
+    replica_list = put_start_result.value();
+    EXPECT_EQ(replica_list.size(), kReplicaCnt);
+    for (size_t i = 0; i < kReplicaCnt; i++) {
+        EXPECT_EQ(ReplicaStatus::PROCESSING, replica_list[i].status);
+    }
+
+    // Wait for a while until key_2 can be discarded and released.
+    for (size_t i = 0; i <= master_config.put_start_release_timeout_sec; i++) {
+        for (auto& context : contexts) {
+            auto result = service_->Ping(context.client_id);
+            EXPECT_TRUE(result.has_value());
+        }
+        // Protect key_1 from eviction.
+        auto get_result = service_->GetReplicaList(key_1, "default");
+        EXPECT_TRUE(get_result.has_value());
+        std::this_thread::sleep_for(std::chrono::seconds(1));
+    }
+
+    // Put key_2 again, should fail because eviction has not been triggered. And
+    // this PutStart should trigger the eviction.
+    put_start_result =
+        service_->PutStart(client_id, key_2, "default", slice_length, config);
+    EXPECT_FALSE(put_start_result.has_value());
+    EXPECT_EQ(put_start_result.error(), ErrorCode::NO_AVAILABLE_HANDLE);
+
+    // Wait a moment for the eviction to complete.
+    std::this_thread::sleep_for(std::chrono::milliseconds(50));
+
+    // Put key_2 again, should success because the previous one has been
+    // discarded and released.
+    put_start_result =
+        service_->PutStart(client_id, key_2, "default", slice_length, config);
+    EXPECT_TRUE(put_start_result.has_value());
+    replica_list = put_start_result.value();
+    EXPECT_EQ(replica_list.size(), kReplicaCnt);
+    for (size_t i = 0; i < kReplicaCnt; i++) {
+        EXPECT_EQ(ReplicaStatus::PROCESSING, replica_list[i].status);
+    }
+
+    // Complete key_2.
+    put_end_result =
+        service_->PutEnd(client_id, key_2, "default", ReplicaType::MEMORY);
+    EXPECT_TRUE(put_end_result.has_value());
+}
+
+}  // namespace mooncake::test
